requirements.txt
django==4.2.13
djangorestframework==3.16.0
psycopg2-binary
gunicorn